# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenLicensd API
  description: |
    HTTP API for OpenLicensd — an open source license server for creating and validating license keys.

    Admin endpoints require a session cookie obtained from `POST /api/v1/auth/login` or OIDC SSO.
    Unsafe methods also require the `X-CSRF-Token` header matching the `openlicensd_csrf` cookie.
    Public endpoints (`/validate`, `/registry-credentials`, `/auth/login`, OIDC login/callback) do not require authentication.
    Unauthenticated endpoints are rate limited per client IP; configure `OPENLICENSD_TRUSTED_PROXIES` behind a reverse proxy.

    All responses include browser security headers: `Content-Security-Policy`, `X-Frame-Options`,
    and `X-Content-Type-Options`. When `OPENLICENSD_COOKIE_SECURE=true`, responses also include
    `Strict-Transport-Security`.

    The `/api/v1/registry-credentials` endpoint is only available when
    `OPENLICENSD_HARBOR_ENABLED=true` on the server.

    ## Role-based access control

    Authenticated endpoints enforce one of three roles. Insufficient role returns `403` with
    `{"error":"forbidden"}`.

    | Role | Permissions |
    |------|-------------|
    | `admin` | Full access including user management |
    | `operator` | Create, update, and delete licenses, products, and policies |
    | `viewer` | Read-only access to licenses, products, and policies |

    Any authenticated user (any role) may call `/api/v1/auth/me`, `/api/v1/auth/logout`, and
...
  license:
...
servers:
...
paths:
...
  /api/v1/licenses/{id}/machines:
    parameters:
      - $ref: "#/components/parameters/LicenseId"
    get:
      tags: [Licenses]
      summary: List machines for a license
      description: |
        **Required role:** `viewer`, `operator`, or `admin`.

        Returns paginated machine activations for the license. Use `status=active` or
        `status=released` to filter by activation state.
      operationId: listLicenseMachines
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/MachineSort"
        - $ref: "#/components/parameters/Order"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, released]
      responses:
        "200":
          description: Paginated machines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLicenseMachines"
        "400":
          description: Invalid license ID or query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: openlicensd_session
      description: |
        Session cookie obtained from `POST /api/v1/auth/login` or OIDC SSO.
        Unsafe methods (POST, PATCH, DELETE) also require the `X-CSRF-Token` header
        set to the value of the `openlicensd_csrf` cookie.
  parameters:
    LicenseId:
      name: id
      in: path
      required: true
      description: License UUID
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-based)
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: Number of items per page
...
    MachineSort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: [created_at, first_seen_at, last_seen_at, validation_count, fingerprint]
        default: created_at
      description: Sort field for license machines
...
    Order:
      name: order
      in: query
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort direction
  responses:
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            invalid_session:
              value:
                error: invalid session
            unauthorized:
              value:
                error: unauthorized
            missing_csrf:
              value:
                error: missing csrf token
            invalid_csrf:
              value:
                error: invalid csrf token
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            license:
              value:
                error: license not found
            product:
              value:
                error: product not found
            policy:
              value:
                error: policy not found
            user:
              value:
                error: user not found
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
...
    LicenseMachine:
      type: object
      required:
        - id
        - license_id
        - fingerprint
        - display_name
        - first_seen_at
        - last_seen_at
        - validation_count
      properties:
        id:
          type: string
          format: uuid
        license_id:
          type: string
          format: uuid
        fingerprint:
          type: string
        name:
          type: ["string", "null"]
        hostname:
          type: ["string", "null"]
        display_name:
          type: string
        first_seen_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        last_seen_ip:
          type: ["string", "null"]
        validation_count:
          type: integer
          format: int64
        deactivated_at:
          type: ["string", "null"]
          format: date-time
        deactivated_by:
          type: ["string", "null"]
          format: uuid
    PaginatedLicenseMachines:
      type: object
      required: [items, page, page_size, total, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LicenseMachine"
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
          format: int64
        total_pages:
          type: integer
...